# PERF-#5225: do not convert 'value' to a list at '.insert()'